Reorganize stubbed unit test code
