+++ xyris.code-workspace
@@ -23,8 +23,9 @@
 		],
 		"clang-format.fallbackStyle": "Webkit",
 		"C_Cpp.intelliSenseEngineFallback": "Enabled",
+        "C_Cpp.exclusionPolicy": "checkFilesAndFolders",
         "C_Cpp.codeAnalysis.exclude": {
             "${workspaceRoot}/Tests/**": true
-        }
+        },
 	}
 }
